ehu-surveillanceV2-main/ocr-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
paddlepaddle==2.6.0
paddleocr==2.7.3
easyocr==1.7.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-multipart==0.0.6
pydantic==2.6.0
httpx==0.27.0
rapidfuzz==3.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
